requirements.txt
torch==1.11.0
torch-scatter==2.0.9
torch-sparse==0.6.13
torch_geometric==2.5.3
PyYAML==6.0.1
scipy==1.10.1
networkx==3.1
timm==1.0.7
charset-normalizer==3.1.0
matplotlib==3.7.5
ogb==1.3.6